requirements.txt
requests
python-dotenv
pandas
python-docx
streamlit
pytest
openpyxl